backend/auth.py
from passlib.context import CryptContext
from jose import jwt
from datetime import datetime, timedelta

# IMPORTANT: Replace this with a long, random string.
# In a real app, load this from an environment variable.
SECRET_KEY = "5027811d2f73088d1aa1299f95d5ffa351743b9c67c22c2a2d29dfb5313a901c"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Setup password hashing context
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt